worker/requirements.txt
redis
Pillow
boto3